requirements.txt
babel== 2.10.3
python-dateutil==2.7.0
flask-migrate==3.1.0
flask-moment==1.0.4
flask-wtf==1.0.1
flask==2.1.3
flask_sqlalchemy==2.5.1

